requirements.txt
pyTelegramBotAPI==4.14.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytest==7.4.3 